money/requirements.txt
asgiref==3.10.0
attrs==25.4.0
Automat==25.4.16
cffi==2.0.0
click==8.3.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
cssbeautifier==1.15.4
daphne==4.2.1
Django==5.2.8
django-allauth==65.13.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-extensions==4.1
django-htmx==1.26.0
django-stubs==5.2.7
django-stubs-ext==5.2.7
djlint==1.36.4
EditorConfig==0.17.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
jsbeautifier==1.15.4
json5==0.12.1
mysqlclient==2.1.1
pathspec==0.12.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.11.3
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
whitenoise==6.11.0
zope.interface==8.1.1
